infra/ecs/embed_and_sync_css/requirements.txt
psycopg2-binary
requests
sentence-transformers
python-dotenv
